app/requirements.txt
Flask==0.12.2
gunicorn==18.0
python-dotenv==0.7.1
